build(unity-bootstrap-theme): fixed webpack config and moved global-header file
